Refactor service update
