+++ service-model-generator/modules/service-model-generator-ls-extension/src/main/java/io/ballerina/servicemodelgenerator/extension/model/UpdateModelContext.java
@@ -1,0 +1,13 @@
+package io.ballerina.servicemodelgenerator.extension.model;
+
+import io.ballerina.compiler.api.SemanticModel;
+import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
+import io.ballerina.compiler.syntax.tree.ServiceDeclarationNode;
+import io.ballerina.projects.Document;
+import io.ballerina.projects.Project;
+import org.ballerinalang.langserver.commons.workspace.WorkspaceManager;
+
+public record UpdateModelContext(Service service, Function function, SemanticModel semanticModel, Project project,
+                                 WorkspaceManager workspaceManager, String filePath, Document document,
+                                 ServiceDeclarationNode serviceNode, FunctionDefinitionNode functionNode) {
+}
