+++ service-model-generator/modules/service-model-generator-ls-extension/src/main/java/io/ballerina/servicemodelgenerator/extension/service/AiChatServiceBuilder.java
@@ -2,8 +2,6 @@ package io.ballerina.servicemodelgenerator.extension.service;
 
 import io.ballerina.compiler.syntax.tree.ModulePartNode;
 import io.ballerina.servicemodelgenerator.extension.model.AddModelContext;
-import io.ballerina.servicemodelgenerator.extension.model.ModelFromSourceContext;
-import io.ballerina.servicemodelgenerator.extension.model.NodeBuilder;
 import io.ballerina.servicemodelgenerator.extension.model.Service;
 import io.ballerina.servicemodelgenerator.extension.util.ListenerUtil;
 import io.ballerina.servicemodelgenerator.extension.util.Utils;
@@ -19,7 +17,6 @@ import java.util.Optional;
 import java.util.Set;
 
 import static io.ballerina.servicemodelgenerator.extension.ServiceModelGeneratorConstants.NEW_LINE;
-import static io.ballerina.servicemodelgenerator.extension.service.ServiceBuilderRouter.DEFAULT_SERVICE_BUILDER;
 import static io.ballerina.servicemodelgenerator.extension.util.Constants.AI;
 import static io.ballerina.servicemodelgenerator.extension.util.Constants.BALLERINA;
 import static io.ballerina.servicemodelgenerator.extension.util.Constants.HTTP;
@@ -31,10 +28,11 @@ import static io.ballerina.servicemodelgenerator.extension.util.Utils.getService
 import static io.ballerina.servicemodelgenerator.extension.util.Utils.importExists;
 import static io.ballerina.servicemodelgenerator.extension.util.Utils.populateRequiredFuncsDesignApproachAndServiceType;
 
-public class AiChatServiceBuilder implements NodeBuilder<Service> {
+public final class AiChatServiceBuilder extends AbstractServiceBuilder {
+
     @Override
     public Optional<Service> getModelTemplate(String moduleName) {
-        return DEFAULT_SERVICE_BUILDER.getModelTemplate(moduleName);
+        return super.getModelTemplate(moduleName);
     }
 
     @Override
@@ -78,16 +76,6 @@ public class AiChatServiceBuilder implements NodeBuilder<Service> {
         }
 
         return Map.of(context.filePath(), edits);
-    }
-
-    @Override
-    public Map<String, List<TextEdit>> updateModel(Service model, String filePath) {
-        return DEFAULT_SERVICE_BUILDER.updateModel(model, filePath);
-    }
-
-    @Override
-    public Service getModelFromSource(ModelFromSourceContext context) {
-        return DEFAULT_SERVICE_BUILDER.getModelFromSource(context);
     }
 
     @Override
